Updated Joint class with Victor drive code.

+++ source/joint.h
@@ -6,7 +6,8 @@
 class Joint
 {
 	public:
-		Joint(int pwmPin)
+		Joint(short pwmPin)
+
 }
 
 #endif
